refactor: 테스트 패키지 구조를 프로덕션 패키지 구조와 동일하게 변경
