+++ src/test/java/roomescape/time/domain/ReservationTimeTest.java
@@ -1,9 +1,8 @@
-package roomescape.admin.domain;
+package roomescape.time.domain;
 
 import java.time.LocalTime;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import roomescape.time.domain.ReservationTime;
 
 class ReservationTimeTest {
