+++ src/test/java/roomescape/time/repository/H2ReservationTimeRepositoryTest.java
@@ -1,4 +1,4 @@
-package roomescape.admin.repository.time;
+package roomescape.time.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
